UPSTREAM: Staging: iio: tsl2563 ambient light sensor driver
Add driver support for the tsl2563 TAOS ambient light sensor. After looking at
discussions on LKML, the driver was modified from a 'hwmon' driver to an 'iio'
driver. The sysfs interfaces have been tested on an RX51 (N900) to see if it
responds to changing light conditions.

The only real reason for submitting this to staging is that it is dependent on
the IIO subsystem.

Signed-off-by: Amit Kucheria <amit.kucheria@verdurent.com>
Cc: Jonathan Cameron <jic23@cam.ac.uk>
Cc: linux-omap@vger.kernel.org
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>

(cherry picked from commit ee1f1fa4071416b130fd0771d1f26ea3f41e1fc3)

Signed-off-by: Mandeep Singh Baines <msb@chromium.org>