UBUNTU: [Config] Disable CONFIG_USB_DEVICEFS
BugLink: http://bugs.launchpad.net/bugs/488274

This option is incompatible with the udev in karmic and lucid, leading
to failure to load firmware files for various pieces of USB kit. It is
also now deprecated and going away very shortly. We should therefore
re-disable it. More details are in the thread below:

    http://markmail.org/message/3mw5yw465qmxgnwp

Although older versions of Virtualbox and VMWare needed it we do not
believe that current versions use it.

This reverts the change done requested for Karmic by the following
bug report:

    https://bugs.launchpad.net/bugs/417748

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>